Fix the RSS feed reposting (#619)
When we stopped including the date in the file we broke the date in RSS
feeds and made it so the posts had the build time as the publish time.
We need to specify the post date from now on which is what they do in
their own docu release note blog posts.

Signed-off-by: Tim Smith <tsmith84@gmail.com>