Update overview.md (#756)
I like to follow a best practice of never starting a sentence with "it". "It" forces the reader to think about what "it" is referring to. In this case, the "it" is not unclear, but I still like to make a habit of writing with a style of clarity and less cognitive load for the reader. C'est tout.